refactor commands into separate directories, continue work on toggle-able commands
