+++ assets/src/components/content/StructuredContentEditor.tsx
@@ -3,7 +3,7 @@ import { Node } from 'slate';
 import { StructuredContent } from 'data/content/resource';
 import { Selection } from 'data/content/model';
 import { Editor } from 'components/editor/Editor';
-import { ToolbarItem } from 'components/resource/toolbar';
+import { ToolbarItem } from 'components/editor/toolbars/toolbar';
 import { ProjectSlug } from 'data/types';
 import { ErrorBoundary } from 'components/common/ErrorBoundary';
 import { ToolbarPosition } from 'components/editor/toolbars/Fixed';
