+++ assets/src/components/editor/commands/buttons/Table.tsx
@@ -1,0 +1,42 @@
+import { ReactEditor } from 'slate-react';
+import { SizePicker } from 'components/editor/toolbars/buttons/SizePicker';
+import { CommandDesc, Command } from 'components/editor/toolbars/interfaces';
+import { Transforms } from 'slate';
+import { td, tr, table } from 'data/content/model';
+
+// The UI command for creating tables
+const command: Command = {
+  execute: (context: any, editor: ReactEditor, params: any) => {
+
+    const rows: any = [];
+
+    for (let i = 0; i < params.rows; i += 1) {
+      const tds = [];
+      for (let j = 0; j < params.columns; j += 1) {
+        tds.push(td(''));
+      }
+      rows.push(tr(tds));
+    }
+
+    const t = table(rows);
+    Transforms.insertNodes(editor, t);
+  },
+  precondition: (editor: ReactEditor) => {
+
+    return true;
+  },
+
+  obtainParameters: (editor: ReactEditor,
+    onDone: (params: any) => void, onCancel: () => void) => {
+
+    return <SizePicker onHide={onCancel}
+      onTableCreate={(rows, columns) => onDone({ rows, columns })} />;
+  },
+};
+
+export const commandDesc: CommandDesc = {
+  type: 'CommandDesc',
+  icon: 'grid_on',
+  description: 'Table',
+  command,
+};
