+++ assets/src/components/editor/editors/Audio.tsx
@@ -3,70 +3,9 @@ import { ReactEditor } from 'slate-react';
 import { Transforms } from 'slate';
 import { updateModel, getEditMode } from './utils';
 import * as ContentModel from 'data/content/model';
-import { Command, CommandDesc } from '../interfaces';
 import { EditorProps, CommandContext } from './interfaces';
-import guid from 'utils/guid';
 import * as Settings from './Settings';
-import { MIMETYPE_FILTERS, SELECTION_TYPES } from 'components/media/manager/MediaManager';
-import ModalSelection from 'components/modal/ModalSelection';
-import { MediaManager } from 'components/media/manager/MediaManager.controller';
-import { modalActions } from 'actions/modal';
-import { MediaItem } from 'types/media';
-
-const dismiss = () => (window as any).oliDispatch(modalActions.dismiss());
-const display = (c: any) => (window as any).oliDispatch(modalActions.display(c));
-
-export function selectAudio(projectSlug: string,
-  model: ContentModel.Audio): Promise<ContentModel.Audio> {
-
-  return new Promise((resolve, reject) => {
-
-    const selected = { img: null };
-
-    const mediaLibrary =
-      <ModalSelection title="Select audio"
-        onInsert={() => { dismiss(); resolve(selected.img as any); }}
-        onCancel={() => dismiss()}
-        disableInsert={true}
-      >
-        <MediaManager model={model}
-          projectSlug={projectSlug}
-          onEdit={() => { }}
-          mimeFilter={MIMETYPE_FILTERS.AUDIO}
-          selectionType={SELECTION_TYPES.SINGLE}
-          initialSelectionPaths={[model.src]}
-          onSelectionChange={(images: MediaItem[]) => {
-            const first : ContentModel.Audio = { type: 'audio', src: images[0].url,
-              children: [{ text: '' }], id: guid() };
-            (selected as any).img = first;
-          }} />
-      </ModalSelection>;
-
-    display(mediaLibrary);
-  });
-}
-
-const command: Command = {
-  execute: (context, editor: ReactEditor) => {
-    const audio = ContentModel.create<ContentModel.Audio>(
-      { type: 'audio', src: '', children: [{ text: '' }], id: guid() });
-    selectAudio(context.projectSlug, audio)
-    .then((img) => {
-      Transforms.insertNodes(editor, img);
-    });
-  },
-  precondition: (editor: ReactEditor) => {
-
-    return true;
-  },
-};
-
-export const commandDesc: CommandDesc = {
-  type: 'CommandDesc',
-  icon: 'audiotrack',
-  description: 'Audio Clip',
-  command,
-};
+import { selectAudio } from '../toolbars/buttons/Audio';
 
 
 type AudioSettingsProps = {
