+++ assets/src/components/editor/editors/Image.tsx
@@ -3,69 +3,7 @@ import { ReactEditor, useFocused, useSelected } from 'slate-react';
 import { Transforms } from 'slate';
 import { updateModel } from './utils';
 import * as ContentModel from 'data/content/model';
-import { Command, CommandDesc } from '../interfaces';
 import { EditorProps } from './interfaces';
-import guid from 'utils/guid';
-import { MIMETYPE_FILTERS, SELECTION_TYPES } from 'components/media/manager/MediaManager';
-import ModalSelection from 'components/modal/ModalSelection';
-import { MediaManager } from 'components/media/manager/MediaManager.controller';
-import { modalActions } from 'actions/modal';
-import { MediaItem } from 'types/media';
-
-const dismiss = () => (window as any).oliDispatch(modalActions.dismiss());
-const display = (c: any) => (window as any).oliDispatch(modalActions.display(c));
-
-export function selectImage(projectSlug: string,
-  model: ContentModel.Image): Promise<ContentModel.Image> {
-
-  return new Promise((resolve, reject) => {
-
-    const selected = { img: null };
-
-    const mediaLibrary =
-        <ModalSelection title="Select an image"
-          onInsert={() => { dismiss(); resolve(selected.img as any); }}
-          onCancel={() => dismiss()}
-          disableInsert={true}
-        >
-          <MediaManager model={model}
-            projectSlug={projectSlug}
-            onEdit={() => { }}
-            mimeFilter={MIMETYPE_FILTERS.IMAGE}
-            selectionType={SELECTION_TYPES.SINGLE}
-            initialSelectionPaths={model.src ? [model.src] : [selected.img as any]}
-            onSelectionChange={(images: MediaItem[]) => {
-              const first : ContentModel.Image = { type: 'img', src: images[0].url,
-                children: [{ text: '' }], id: guid()};
-              (selected as any).img = first;
-            }} />
-        </ModalSelection>;
-
-    display(mediaLibrary);
-  });
-}
-
-const command: Command = {
-  execute: (context, editor: ReactEditor) => {
-    const image = ContentModel.create<ContentModel.Image>(
-      { type: 'img', src: '', children: [{ text: '' }], id: guid() });
-    selectImage(context.projectSlug, image)
-    .then((img) => {
-      Transforms.insertNodes(editor, img);
-    });
-  },
-  precondition: (editor: ReactEditor) => {
-
-    return true;
-  },
-};
-
-export const commandDesc: CommandDesc = {
-  type: 'CommandDesc',
-  icon: 'image',
-  description: 'Image',
-  command,
-};
 
 export interface ImageProps extends EditorProps<ContentModel.Image> {
 }
@@ -89,7 +27,7 @@ export const ImageEditor = (props: ImageProps) => {
     onEdit(Object.assign({}, model, { caption: text, alt: text }));
 
   const imageStyle = focused && selected
-  ? { border: 'solid 2px lightblue' } : {};
+  ? { border: 'solid 3px lightblue', borderRadius: 0 } : { border: 'solid 3px transparent' };
 
   // Note that it is important that any interactive portions of a void editor
   // must be enclosed inside of a "contentEditable=false" container. Otherwise,
@@ -104,18 +42,17 @@ export const ImageEditor = (props: ImageProps) => {
           <figure>
             <img
               style={imageStyle}
-              className="img-fluid img-thumbnail"
+              className="img-fluid"
               src={model.src}
               draggable={false}
+              onClick={e => Transforms.select(editor, ReactEditor.findPath(editor, model))}
             />
             <figcaption>
               <input
                 type="text"
                 value={model.caption}
                 placeholder="Type caption for image"
-                onChange={e => {
-                  setCaptionAndAlt(e.target.value)
-                }}
+                onChange={e => setCaptionAndAlt(e.target.value)}
                 // onKeyPress={e => e.key === 'Enter' ?  : null}
                 // onKeyPress={e => Settings.onEnterApply(e, () => onEdit(model))}
                 className="caption-editor"
