+++ assets/src/components/editor/editors/Link.tsx
@@ -2,14 +2,11 @@ import React, { useState, useEffect, useRef } from 'react';
 import Popover from 'react-tiny-popover';
 import * as ContentModel from 'data/content/model';
 import { ReactEditor } from 'slate-react';
-import { Transforms, Node, Range, Path, Editor, Text } from 'slate';
+import { Transforms } from 'slate';
 import { EditorProps, CommandContext } from './interfaces';
-import { Command, CommandDesc } from '../interfaces';
 import { updateModel } from './utils';
 import { Action, onEnterApply } from './Settings';
 import * as Persistence from 'data/persistence/resource';
-
-import guid from 'utils/guid';
 
 const internalLinkPrefix = '/course/link';
 
@@ -24,7 +21,7 @@ const addProtocol = (href: string) => isValidHref(href)
   ? href
   : 'http://' + href;
 
-const normalizeHref = (href: string) => addProtocol(href.trim());
+export const normalizeHref = (href: string) => addProtocol(href.trim());
 
 // Takes a delivery oriented internal link and translates it to
 // a link that will resolve at authoring time. This allows
@@ -32,85 +29,6 @@ const normalizeHref = (href: string) => addProtocol(href.trim());
 // page.
 const translateDeliveryToAuthoring = (href: string, projectSlug: string) => {
   return `/project/${projectSlug}/resource/` + href.substr(href.lastIndexOf('/') + 1);
-};
-
-const command: Command = {
-  execute: (context, editor: ReactEditor) => {
-
-    const selection = editor.selection;
-    if (selection === null) return;
-
-    // HELPERS
-
-    // make sure we arent creating a link within a link
-    const isLinkPresent = () => {
-
-      // const parentNodeHasLink = () => {
-      //   const parentType = Node.parent(editor, editor.selection?.anchor.path as Path)?.type;
-      //   return parentType === 'a';
-      // };
-
-      const childNodesHaveLink = () => {
-        return Node.fragment(editor, selection)
-          .map(node => Array.from(Node.descendants(node))
-            .reduce((acc, [node]) => node.type === 'a' ? true : acc, false))
-          .some(bool => bool);
-      };
-      // parentNodeHasLink()
-      return childNodesHaveLink();
-    };
-
-    const wrapLink = (editor: ReactEditor, range: Range, link: ContentModel.Hyperlink) => {
-      Transforms.wrapNodes(editor, link, { split: true, at: range });
-      Transforms.collapse(editor, { edge: 'end' });
-    };
-
-    // Wrap the selection with a HyperLink
-    const addLink = () => {
-      const offset1 = selection.anchor.offset < selection.focus.offset
-        ? selection.anchor.offset : selection.focus.offset;
-      const offset2 = offset1 === selection.anchor.offset
-        ? selection.focus.offset : selection.anchor.offset;
-
-      const [node] = Editor.node(editor,
-        Editor.range(editor, selection.anchor, selection.focus));
-
-      const hrefText = Text.isText(node) ? node.text.slice(offset1, offset2) : 'Link URL';
-
-      const newLink = ContentModel.create<ContentModel.Hyperlink>({
-        type: 'a', href: normalizeHref(hrefText), target: 'self',
-        children: [{ text: '' }], id: guid(),
-      });
-
-      wrapLink(editor, selection, newLink);
-    };
-
-    const removeLinks = (editor: ReactEditor, range: Range) => {
-      // const linksInRange =
-
-      Transforms.unwrapNodes(editor, { at: range, match: node => node.type === 'a' });
-      Transforms.collapse(editor, { edge: 'end' });
-    };
-
-    // LOGIC
-
-    if (isLinkPresent()) {
-      return removeLinks(editor, selection);
-    }
-
-    return addLink();
-  },
-  precondition: (editor: ReactEditor) => {
-    return true;
-  },
-};
-
-
-export const commandDesc: CommandDesc = {
-  type: 'CommandDesc',
-  icon: 'insert_link',
-  description: 'Link',
-  command,
 };
 
 type ExistingLinkProps = {
