+++ assets/src/components/editor/editors/Lists.tsx
@@ -1,100 +1,6 @@
-import { ReactEditor } from 'slate-react';
 import { Transforms, Range, Node, Point, Path, Editor as SlateEditor } from 'slate';
 import * as ContentModel from 'data/content/model';
-import { Command, CommandDesc, CommandContext } from '../interfaces';
-import guid from 'utils/guid';
 import { KeyboardEvent } from 'react';
-import { getNearestBlock, getNearestTopLevel } from '../utils';
-
-const li = () => ContentModel.create<ContentModel.ListItem>(
-  { type: 'li', children: [{ text: '' }], id: guid() });
-
-const ol = () => ContentModel.create<ContentModel.OrderedList>(
-  { type: 'ol', children: [li()], id: guid() });
-
-const ul = () => ContentModel.create<ContentModel.UnorderedList>(
-  { type: 'ul', children: [li()], id: guid() });
-
-
-const toggleList = (editor: ReactEditor, listType: string) => {
-
-  try {
-
-    // The edits here result in intermediate states that normalization
-    // would seek to correct.  So to allow this operation to succeed,
-    // we instruct our editor instance to suspend normalization.
-    (editor as any).suspendNormalization = true;
-
-    const isActive = isActiveList(editor);
-
-    Transforms.unwrapNodes(editor, {
-      match: n => n.type === 'ul' || n.type === 'ol',
-      split: true,
-    });
-
-    Transforms.setNodes(editor, {
-      type: isActive ? 'p' : 'li',
-    });
-
-    if (!isActive) {
-      const block = { type: listType, children: [] };
-      Transforms.wrapNodes(editor, block);
-    }
-  } catch (error) {
-    // tslint:disable-next-line
-    console.error(error);
-
-  } finally {
-    // Whether the operation succeeded or failed, we restore
-    // normalization
-    (editor as any).suspendNormalization = false;
-  }
-
-};
-
-const isActiveList = (editor: ReactEditor) => {
-  const [match] = SlateEditor.nodes(editor, {
-    match: n => n.type === 'ul' || n.type === 'ol',
-  });
-
-  return !!match;
-};
-
-const listCommandMaker = (listType: string) => {
-  return {
-    execute: (context: CommandContext, editor: ReactEditor) => {
-      toggleList(editor, listType);
-    },
-    precondition: (editor: ReactEditor) => {
-      return true;
-    },
-  };
-};
-
-const ulCommand: Command = listCommandMaker('ul');
-const olCommand: Command = listCommandMaker('ol');
-
-export const ulCommandDesc: CommandDesc = {
-  type: 'CommandDesc',
-  icon: 'format_list_bulleted',
-  description: 'Unordered List',
-  command: ulCommand,
-  active: (editor: ReactEditor) => getNearestTopLevel(editor).caseOf({
-    just: n => n.type === 'ul',
-    nothing: () => false,
-  }),
-};
-
-export const olCommandDesc: CommandDesc = {
-  type: 'CommandDesc',
-  icon: 'format_list_numbered',
-  description: 'Ordered List',
-  command: olCommand,
-  active: (editor: ReactEditor) => getNearestTopLevel(editor).caseOf({
-    just: n => n.type === 'ol',
-    nothing: () => false,
-  }),
-};
 
 const isList = (n: Node) => n.type === 'ul' || n.type === 'ol';
 
@@ -126,7 +32,7 @@ function handleIndent(editor: SlateEditor, e: KeyboardEvent) {
               const item = parent.children[i];
               if (isList(item)) {
 
-                const newList = item.type === 'ul' ? ul() : ol();
+                const newList = item.type === 'ul' ? ContentModel.ul() : ContentModel.ol();
                 newList.children.pop();
 
                 Transforms.wrapNodes(editor, newList, { at: editor.selection });
@@ -137,7 +43,7 @@ function handleIndent(editor: SlateEditor, e: KeyboardEvent) {
           }
 
           // Allow indent with the same list type as current parent
-          const newList = parent.type === 'ul' ? ul() : ol();
+          const newList = parent.type === 'ul' ? ContentModel.ul() : ContentModel.ol();
           newList.children.pop();
 
           Transforms.wrapNodes(editor, newList, { at: editor.selection });
@@ -211,11 +117,8 @@ function handleTermination(editor: SlateEditor, e: KeyboardEvent) {
           // outside of the parent list
           Transforms.removeNodes(editor, { at: path });
 
-          const p = ContentModel.create<ContentModel.Paragraph>(
-            { type: 'p', children: [{ text: '' }], id: guid() + ''  });
-
           // Insert it ahead of the next node
-          Transforms.insertNodes(editor, p, { at: Path.next(parentPath) });
+          Transforms.insertNodes(editor, ContentModel.p(), { at: Path.next(parentPath) });
           Transforms.select(editor, Path.next(parentPath));
 
           e.preventDefault();
