+++ assets/src/components/resource/Editors.scss
@@ -17,6 +17,7 @@
 
       .slate-editor {
         padding: 1.3em;
+        padding-left: 42px;
       }
     }
