+++ assets/src/components/resource/Editors.tsx
@@ -7,7 +7,7 @@ import {
 } from 'data/content/resource';
 import { ActivityEditorMap, EditorDesc } from 'data/content/editors';
 import { UnsupportedActivity } from './UnsupportedActivity';
-import { getToolbarForResourceType } from './toolbar';
+import { getToolbarForResourceType } from '../editor/toolbars/toolbar';
 import { StructuredContentEditor } from 'components/content/StructuredContentEditor';
 import { ProjectSlug, ResourceSlug } from 'data/types';
 import { AbbreviatedActivity } from './AbbreviatedActivity';
