+++ assets/styles/authoring/layout/workspace.scss
@@ -19,7 +19,7 @@
     padding-bottom: 40px;
 
     main {
-      padding: 20px;
+      padding: 32px;
     }
   }
 
@@ -27,7 +27,7 @@
     width: 100%;
     border-bottom: 1px solid $gray-400;
 
-    padding-left: 24px;
+    padding-left: 32px;
     padding-top: 12px;
     padding-bottom: 6px;
     padding-right: 12px;
