+++ assets/styles/authoring/main.scss
@@ -39,6 +39,13 @@ body {
 
 }
 
+.container {
+  margin-left: 0;
+  margin-right: 0;
+  padding-left: 0;
+  padding-right: 0;
+}
+
 .oli-logo {
   font-family: 'Open Sans', sans-serif;
 }
